refactor(lyrics): split orchestrator.rs tests into sibling files
orchestrator.rs was at 967/1000 lines after the V19 source-gating
work landed (#102). Extracted both test modules to sibling files via
#[path] attribute (same pattern as PR #106 worker_tests/mod_tests):

- orchestrator_gate_tests.rs — `is_allowed_text_source` unit tests (~113 lines)
- orchestrator_tests.rs       — `Orchestrator` integration tests (~440 lines)

orchestrator.rs is now 423 lines. No behavior change; existing tests
run unchanged from the new locations.

Closes #104
